Extract wire.rs
